A little cleanup and organizing
